grid in work

+++ wizard-ejb/src/main/java/com/engagepoint/labs/wizard/questions/RuleExecutor.java
@@ -119,7 +119,7 @@ public class RuleExecutor implements Serializable {
     }
 
     private boolean compareFile(Value parentQuestionAnswer, String stringToCompareWith) {
-        return ((parentQuestionAnswer.getValue()) != 0) && (stringToCompareWith.equals("true"));
+        return ((Integer)(parentQuestionAnswer.getValue()) != 0) && (stringToCompareWith.equals("true"));
     }
 
     private boolean compareList(Value parentQuestionAnswer, String[] stringArrayToCompareWith) {
